rename: Termscheck 폴더 이동 login -> join
