+++ src/components/join/Termscheck.jsx
